// src/interfaces/Car.ts
export interface Car {
  id: string;
  plate: string;
  manufacture: string;
  model: string;
  image: string;
  rent_per_day: number;
  capacity: number;
  description: string;
  available_at: string;
  transmission: string;
  available: boolean;
  type: string;
  year: number;
  options: string;
  specs: string;
  created_by: string;
  updated_by: string;
  created_at: string;
  updated_at: string;
}